add test and a couple of fix
resolving virtual method for mark and compact.
Allow instanciation of MarkAndCompact.
Begining Tests

+++ src/vm/GarbageCollector/MarkAndCompactGC.h
@@ -46,7 +46,7 @@ public:
 	void allocateWeakContainersArray();
 	void allocateEphemeronsArray();
 	void forgetNativeObjects();
-	void followCountStartingAt(slot_t *frame, ulong size, ulong startIndex);
+	void followCountStartingAt(slot_t *frame, int size, long startIndex);
 	void setNewPositions(GCSpace * space);
 	void compact(GCSpace * space);
 	void librariesDo();
